Rename logging_utils to output_utils
