+++ hi-ml-histopathology/src/histopathology/utils/output_utils.py
